requirements_mac.txt
accelerate==1.10.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi
charset-normalizer==3.4.4
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.123.2
filelock==3.19.1
fonttools
fsspec
ftfy==6.3.1
greenlet
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools
httpx==0.28.1
hub-sdk==0.0.24
huggingface-hub==0.36.0
idna==3.11
importlib_resources
Jinja2==3.1.6
joblib==1.5.2
kiwisolver
lxml
MarkupSafe
matplotlib
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.8.1.78
opencv-python-headless==4.7.0.72
pandas
pillow
playwright==1.56.0
polars==1.36.0
pydantic==2.12.5
pydantic_core
pyee==13.0.0
py-cpuinfo==9.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz
PyYAML
regex
requests==2.32.5
safetensors==0.7.0
scikit-learn
scipy
seaborn==0.13.2
sentence-transformers==3.1.0
sentencepiece
sentry-sdk==2.47.0
shellingham==1.5.4
soupsieve==2.8
starlette==0.49.3
sympy
thop
threadpoolctl==3.6.0
tokenizers
tqdm==4.67.1
transformers==4.44.2
typing-inspection==0.4.2
typing_extensions
tzdata
ultralytics==8.3.235
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
